protecting contacts routes

+++ controllers/contact.controller.js
@@ -5,7 +5,8 @@ const Contact = require("../models/contact.model")
 // @route GET /api/contacts
 // @access private
 const getContacts = asynHandler(async (req, res) => {
-  const contacts = await Contact.find()
+  const user = req.user
+  const contacts = await Contact.find({ userId: user.id })
   res.json({ contacts })
 })
 
@@ -14,7 +15,8 @@ const getContacts = asynHandler(async (req, res) => {
 // @access private
 const getContact = asynHandler(async (req, res) => {
   const { id } = req.params
-  const contact = await Contact.findById(id)
+  const user = req.user
+  const contact = await Contact.findOne({ userId: user.id, _id: id })
 
   if (!contact) {
     res.status(404)
@@ -29,13 +31,14 @@ const getContact = asynHandler(async (req, res) => {
 // @access private
 const createContact = asynHandler(async (req, res) => {
   const { name, email, phone } = req.body
+  const user = req.user
 
   if (!name || !email || !phone) {
     res.status(400)
     throw new Error("All field are required")
   }
 
-  const contact = await Contact.create({ name, email, phone })
+  const contact = await Contact.create({ name, email, phone, userId: user.id })
 
   res.status(201).json({ contact })
 })
@@ -45,11 +48,17 @@ const createContact = asynHandler(async (req, res) => {
 // @access private
 const updateContact = asynHandler(async (req, res) => {
   const { id } = req.params
+  const user = req.user
 
   const contact = await Contact.findById(id)
   if (!contact) {
     res.status(404)
     throw new Error("Contact not found")
+  }
+
+  if (contact.userId.toString() !== user.id) {
+    res.status(403)
+    throw new Error("Dont have permission to update other user contact")
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(id, req.body, {
@@ -64,11 +73,17 @@ const updateContact = asynHandler(async (req, res) => {
 // @access private
 const deleteContact = asynHandler(async (req, res) => {
   const { id } = req.params
+  const user = req.user
 
   const contact = await Contact.findById(id)
   if (!contact) {
     res.status(404)
     throw new Error("Contact not found")
+  }
+
+  if (contact.userId.toString() !== user.id) {
+    res.status(403)
+    throw new Error("Dont have permission to delete other user contact")
   }
 
   await Contact.deleteOne({ _id: id })
